// game-backend/src/auth/modules/permissions/enums/permission-type.enum.ts
export enum PermissionType {
  USERS = 'USERS',
  PERMISSIONS = 'PERMISSIONS',
  ROLES = 'ROLES',
  GAME_COLLECTIONS = 'GAME_COLLECTIONS',
  DATA_MANAGEMENT = 'DATA_MANAGEMENT',
  COLLECTION_SUPERUSER = 'COLLECTION_SUPERUSER',
  SYSTEM_COLLECTION = 'SYSTEM_COLLECTION',
  ADMIN_PANEL = 'ADMIN_PANEL',
}